refactor(storage): share N-Quads text parser
